switch to recycler view and change movie id to in

+++ app/src/main/java/com/jasminelawrence/popularmoviesstage2/MainActivity.java
@@ -189,7 +189,7 @@ public class MainActivity extends AppCompatActivity {
                 String poster_url = "http://image.tmdb.org/t/p/w185/" + poster_path;
 
                 double user_rating = currentMovie.getDouble("vote_average");
-                double id = currentMovie.getDouble("id");
+                int id = currentMovie.getInt("id");
 
                 // Create a new Movie object with the infromation
                 Movie movie = new Movie(id, original_title, poster_url, plot_synopsis, user_rating, release_date, null, null);
